[Cashtab] Make sure we use bigint for nanosats
Summary: These values should always be ints and they are pretty commonly higher than javascript max safe int. So we should be using bigint.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19155

+++ cashtab/src/validation/index.ts
@@ -1244,7 +1244,9 @@ export const getAgoraPartialListPriceError = (
     // this is the unit agora takes, 1 nanosat per 1 tokens at is the min
     const priceNanoSatsPerDecimalizedToken = xecToNanoSatoshis(priceXec);
 
-    if (priceNanoSatsPerDecimalizedToken < Math.pow(10, tokenDecimals)) {
+    if (
+        priceNanoSatsPerDecimalizedToken < BigInt(Math.pow(10, tokenDecimals))
+    ) {
         return 'Price cannot be lower than 1 nanosatoshi per 1 token satoshi';
     }
